+++ src/Geoapify.DependencyInjection/Setup.cs
@@ -1,13 +1,24 @@
 ﻿using System.Net;
+using Geoapify.DependencyInjection;
 using Geoapify.SDK.Client;
 using Geoapify.SDK.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 // ReSharper disable once CheckNamespace
-public static class GeoapifySetup
+public static class Setup
 {
-	public static IServiceCollection AddGeoapify(this IServiceCollection services, string apiKey)
+	public static GeoapifyServiceCollection AddGeoapify(this IServiceCollection services, string apiKey)
 	{
+		if (string.IsNullOrWhiteSpace(apiKey))
+		{
+			throw new ArgumentNullException(nameof(apiKey), "Missing apiKey");
+		}
+
+		if (services.Any(d => d.ImplementationType == typeof(GeoapifyClient)))
+		{
+			throw new InvalidOperationException("A Geoapify client is already registered.");
+		}
+
 		services.AddHttpClient<GeoapifyClient>(client =>
 		{
 			client.BaseAddress = new Uri("https://api.geoapify.com/v1/");
@@ -22,7 +33,8 @@ public static class GeoapifySetup
 		{
 			options.ApiKey = apiKey;
 		});
+		services.AddSingleton(TimeProvider.System);
 		services.AddTransient<IGeoapifyClient, GeoapifyClient>();
-		return services;
+		return new GeoapifyServiceCollection(services);
 	}
 }