+++ tests/Geoapify.IntegrationTests/Configuration/ContainerFixture.cs
@@ -1,12 +1,29 @@
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using MongoDB.Driver;
+using Testcontainers.MongoDb;
 
 namespace Geoapify.IntegrationTests.Configuration;
 
-public class ContainerFixture
+public class ContainerFixture : IAsyncLifetime
 {
+	private readonly MongoDbContainer _mongoContainer;
+
 	public ContainerFixture()
 	{
+		_mongoContainer = new MongoDbBuilder()
+			.WithImage("mongo:latest")
+			.WithUsername("mongo")
+			.WithPassword("secret")
+			.Build();
+	}
+
+	public ServiceProvider Provider { get; private set; } = null!;
+
+	public async ValueTask InitializeAsync()
+	{
+		await _mongoContainer.StartAsync();
+
 		var services = new ServiceCollection();
 
 		var configuration = new ConfigurationBuilder()
@@ -16,9 +33,16 @@ public class ContainerFixture
 
 		var apiKey = configuration["ApiKey"] ?? throw new InvalidOperationException("Missing configuration value: ApiKey");
 
-		services.AddGeoapify(apiKey);
+		var client = new MongoClient(_mongoContainer.GetConnectionString());
+		var db = client.GetDatabase("db");
+		services.AddGeoapify(apiKey)
+			.AddMongoDBStorage(db, "addresses");
 		Provider = services.BuildServiceProvider();
 	}
 
-	public ServiceProvider Provider { get; }
+	public async ValueTask DisposeAsync()
+	{
+		await _mongoContainer.DisposeAsync();
+		await Provider.DisposeAsync();
+	}
 }