+++ tests/Geoapify.UnitTests/GlobalUsings.cs
@@ -1,0 +1,6 @@
+// Global using directives
+
+global using Microsoft.Extensions.DependencyInjection;
+global using Microsoft.Extensions.Hosting;
+global using Microsoft.Extensions.Options;
+global using NSubstitute;